LayoutTests/platform/mac/imported/w3c/web-platform-tests/mathml/relations/css-styling/size-containment-001.tentative-expected.txt


FAIL intrinsic size of maction with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL inline-size of maction with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL intrinsic size of menclose with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 88.03125
FAIL inline-size of menclose with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 88.03125
FAIL intrinsic size of merror with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL inline-size of merror with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL intrinsic size of mfrac with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL inline-size of mfrac with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
PASS intrinsic size of mi with contain: size; contain-intrinsic-inline-size: 111px;
PASS inline-size of mi with contain: size; contain-intrinsic-inline-size: 111px;
FAIL intrinsic size of mmultiscripts with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 132.03125
FAIL inline-size of mmultiscripts with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 132.03125
PASS intrinsic size of mn with contain: size; contain-intrinsic-inline-size: 111px;
PASS inline-size of mn with contain: size; contain-intrinsic-inline-size: 111px;
PASS intrinsic size of mo with contain: size; contain-intrinsic-inline-size: 111px;
PASS inline-size of mo with contain: size; contain-intrinsic-inline-size: 111px;
FAIL intrinsic size of mover with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL inline-size of mover with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL intrinsic size of mpadded with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL inline-size of mpadded with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL intrinsic size of mphantom with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL inline-size of mphantom with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL intrinsic size of mrow with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL inline-size of mrow with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
PASS intrinsic size of ms with contain: size; contain-intrinsic-inline-size: 111px;
PASS inline-size of ms with contain: size; contain-intrinsic-inline-size: 111px;
FAIL intrinsic size of mstyle with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL inline-size of mstyle with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL intrinsic size of msub with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 132.03125
FAIL inline-size of msub with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 132.03125
FAIL intrinsic size of msubsup with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 132.03125
FAIL inline-size of msubsup with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 132.03125
FAIL intrinsic size of msup with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 132.03125
FAIL inline-size of msup with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 132.03125
PASS intrinsic size of mtext with contain: size; contain-intrinsic-inline-size: 111px;
PASS inline-size of mtext with contain: size; contain-intrinsic-inline-size: 111px;
FAIL intrinsic size of munder with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL inline-size of munder with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL intrinsic size of munderover with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL inline-size of munderover with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL intrinsic size of semantics with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
FAIL inline-size of semantics with contain: size; contain-intrinsic-inline-size: 111px; assert_equals: expected 111 but got 76.875
